--- piwheels/audit/__init__.py
@@ -1,52 +1,0 @@
-# The piwheels project
-#   Copyright (c) 2017 Ben Nuttall <https://github.com/bennuttall>
-#   Copyright (c) 2017 Dave Jones <dave@waveform.org.uk>
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#     * Redistributions of source code must retain the above copyright
-#       notice, this list of conditions and the following disclaimer.
-#     * Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in the
-#       documentation and/or other materials provided with the distribution.
-#     * Neither the name of the copyright holder nor the
-#       names of its contributors may be used to endorse or promote products
-#       derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-"""
-Contains the functions that implement the :program:`piw-audit` script.
-
-.. autofunction:: main
-"""
-
-import logging
-
-
-def report(file, prefix, label, path):
-    logging.error('%s %s %s', prefix, label, path)
-    if file:
-        file.write(str(path))
-        file.write('\n')
-
-def report_missing(config, label, path):
-    report(config.missing_file, 'missing', label, path)
-
-def report_extra(config, label, path):
-    report(config.extra_file, 'extraneous', label, path)
-
-def report_broken(config, label, path):
-    report(config.broken_file, 'corrupted', label, path)
-
